Cleanup and improve generation tests.
